Exported servo class to external file
