+++ src/main.py
@@ -6,6 +6,7 @@ import base64
 import RPi.GPIO as GPIO
 from server import startServer
 from time import sleep
+from servo import Servo
 from trajectory import calcFinalAngles
 
 GPIO.setwarnings(False)
@@ -27,48 +28,6 @@ def camera():
 	print(r.content)
 	print("Calculating trajectory...")
 	return calcFinalAngles(((servoY.angle - 2.2) / 0.053))
-
-
-# servo definition with functions
-class Servo():
-	def __init__(self, PIN):
-		self.PIN = PIN
-		GPIO.setup(self.PIN, GPIO.OUT)
-		self.pwm = GPIO.PWM(self.PIN, 50)
-		self.pwm.start(7.5)
-		self.angle = 0
-                self.locked = False
-
-	# set servo to specific angle
-	def setAngle(self, angle):
-		self.angle = 0.053*angle + 2.2
-		self.pwm.ChangeDutyCycle(self.angle)
-
-    # adjust angle of servo
-	def adjustAngle(self, angle):
-		self.setAngle(self.angle + angle)
-
-    # stop servo
-	def stop(self):
-		self.pwm.stop()
-
-    # lock servo and fire
-	def lock(self):
-		self.setAngle(0)
-                self.locked = 1
-
-	# unlock servo to be ready to fire
-	def unlock(self):
-		self.setAngle(180)
-                self.lock = 0
-                self.locked = 0
-
-        def toggleLock(self):
-                if (self.locked):
-                    self.unlock()
-                else:
-                    self.lock()
-
 
 # loop options
 def loop():
